Migrate Cassandra tests to use BaseConnectorTest
Rename TestCassandraIntegrationSmokeTest.java to TestCassandraConnectorTest.java .
Merge BaseCassandraDistributedQueries.java tests to TestCassandraConnectorTest.java.
