--- plugin/trino-cassandra/src/test/java/io/trino/plugin/cassandra/BaseCassandraDistributedQueries.java
@@ -1,149 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.trino.plugin.cassandra;
-
-import io.trino.testing.AbstractTestDistributedQueries;
-import io.trino.testing.MaterializedResult;
-import io.trino.testing.sql.TestTable;
-import org.testng.SkipException;
-
-import java.util.Optional;
-
-import static io.trino.spi.type.VarcharType.VARCHAR;
-import static io.trino.testing.MaterializedResult.resultBuilder;
-import static io.trino.testing.assertions.Assert.assertEquals;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-public abstract class BaseCassandraDistributedQueries
-        extends AbstractTestDistributedQueries
-{
-    @Override
-    protected boolean supportsViews()
-    {
-        return false;
-    }
-
-    @Override
-    protected boolean supportsArrays()
-    {
-        return false;
-    }
-
-    @Override
-    protected boolean supportsCommentOnTable()
-    {
-        return false;
-    }
-
-    @Override
-    protected boolean supportsCommentOnColumn()
-    {
-        return false;
-    }
-
-    @Override
-    public void testCreateSchema()
-    {
-        // Cassandra does not support creating schemas
-    }
-
-    @Override
-    public void testRenameTable()
-    {
-        // Cassandra does not support renaming tables
-    }
-
-    @Override
-    public void testAddColumn()
-    {
-        // Cassandra does not support adding columns
-    }
-
-    @Override
-    public void testRenameColumn()
-    {
-        // Cassandra does not support renaming columns
-    }
-
-    @Override
-    public void testDropColumn()
-    {
-        // Cassandra does not support dropping columns
-    }
-
-    @Override
-    public void testDelete()
-    {
-        assertThatThrownBy(super::testDelete)
-                .hasStackTraceContaining("Delete without primary key or partition key is not supported");
-    }
-
-    @Override
-    public void testShowColumns()
-    {
-        MaterializedResult actual = computeActual("SHOW COLUMNS FROM orders");
-
-        MaterializedResult expectedParametrizedVarchar = resultBuilder(getSession(), VARCHAR, VARCHAR, VARCHAR, VARCHAR)
-                .row("orderkey", "bigint", "", "")
-                .row("custkey", "bigint", "", "")
-                .row("orderstatus", "varchar", "", "")
-                .row("totalprice", "double", "", "")
-                .row("orderdate", "date", "", "")
-                .row("orderpriority", "varchar", "", "")
-                .row("clerk", "varchar", "", "")
-                .row("shippriority", "integer", "", "")
-                .row("comment", "varchar", "", "")
-                .build();
-
-        assertEquals(actual, expectedParametrizedVarchar);
-    }
-
-    @Override
-    protected TestTable createTableWithDefaultColumns()
-    {
-        throw new SkipException("Cassandra connector does not support column default values");
-    }
-
-    @Override
-    protected Optional<DataMappingTestSetup> filterDataMappingSmokeTestData(DataMappingTestSetup dataMappingTestSetup)
-    {
-        String typeName = dataMappingTestSetup.getTrinoTypeName();
-        if (typeName.equals("time")
-                || typeName.equals("timestamp")
-                || typeName.equals("decimal(5,3)")
-                || typeName.equals("decimal(15,3)")
-                || typeName.equals("char(3)")) {
-            // TODO this should either work or fail cleanly
-            return Optional.empty();
-        }
-        return Optional.of(dataMappingTestSetup);
-    }
-
-    @Override
-    protected Optional<DataMappingTestSetup> filterCaseSensitiveDataMappingTestData(DataMappingTestSetup dataMappingTestSetup)
-    {
-        String typeName = dataMappingTestSetup.getTrinoTypeName();
-        if (typeName.equals("char(1)")) {
-            // TODO this should either work or fail cleanly
-            return Optional.empty();
-        }
-        return Optional.of(dataMappingTestSetup);
-    }
-
-    @Override
-    protected String dataMappingTableName(String trinoTypeName)
-    {
-        return "tmp_trino_" + System.nanoTime();
-    }
-}
